requirements.txt
Django==3.2.9
djangorestframework==3.13.1
gunicorn==20.1.0
django-environ==0.8.1
psycopg2-binary==2.8.6
django-rq==2.5.1
requests==2.27.1